# DDD - Falta Print das Telas/Documentação API/Raízes Solidárias.postman_collection.json-OpenApi3Json.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Raízes Solidárias",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    },
    {
      "url": "https://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
...
    "/colheita_voluntario": {
      "get": {
        "tags": [
          "Colheita_Voluntario"
        ],
        "summary": "Listar Colheita_Voluntarios",
        "operationId": "ListarColheita_Voluntarios",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      },
      "post": {
        "tags": [
          "Colheita_Voluntario"
        ],
        "summary": "Cadastrar Colheita_Voluntario",
        "operationId": "CadastrarColheita_Voluntario",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarColheita_VoluntarioRequest"
                  },
                  {
                    "example": {
                      "colheita": {
                        "id_colheita": 1
                      },
                      "voluntario": {
                        "id_usuario": 17
                      }
                    }
                  }
                ]
              },
              "example": {
                "colheita": {
                  "id_colheita": 1
                },
                "voluntario": {
                  "id_usuario": 17
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
    "/colheita_voluntario/1/5/17": {
      "put": {
        "tags": [
          "Colheita_Voluntario"
        ],
        "summary": "Atualizar Colheita_Voluntario",
        "operationId": "AtualizarColheita_Voluntario",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarColheita_VoluntarioRequest"
                  },
                  {
                    "example": {
                      "id_usuario_novo": 17,
                      "id_usuario_antigo": 5,
                      "id_colheita": 1
                    }
                  }
                ]
              },
              "example": {
                "id_usuario_novo": 17,
                "id_usuario_antigo": 5,
                "id_colheita": 1
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
    "/colheita": {
      "get": {
        "tags": [
          "Colheita"
        ],
        "summary": "Listar Colheitas",
        "operationId": "ListarColheitas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      },
      "post": {
        "tags": [
          "Colheita"
        ],
        "summary": "Cadastrar Colheita",
        "operationId": "CadastrarColheita",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarColheitaRequest"
                  },
                  {
                    "example": {
                      "data_colheita": "2023-05-31",
                      "descricao_colheita": "Descrição da colheita"
                    }
                  }
                ]
              },
              "example": {
                "data_colheita": "2023-05-31",
                "descricao_colheita": "Descrição da colheita"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
    "/colheita/12": {
      "put": {
        "tags": [
          "Colheita"
        ],
        "summary": "Atualizar Colheita",
        "operationId": "AtualizarColheita",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarColheitaRequest"
                  },
                  {
                    "example": {
                      "id_colheita": 12,
                      "data_colheita": "2100-05-31",
                      "descricao_colheita": "Descrição da colheita"
                    }
                  }
                ]
              },
              "example": {
                "id_colheita": 12,
                "data_colheita": "2100-05-31",
                "descricao_colheita": "Descrição da colheita"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      },
      "delete": {
        "tags": [
          "Colheita"
        ],
        "summary": "Deletar Colheita",
        "operationId": "DeletarColheita",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "https://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AtualizarColheita_VoluntarioRequest": {
        "title": "AtualizarColheita_VoluntarioRequest",
        "required": [
          "id_usuario_novo",
          "id_usuario_antigo",
          "id_colheita"
        ],
        "type": "object",
        "properties": {
          "id_usuario_novo": {
            "type": "integer",
            "format": "int32"
          },
          "id_usuario_antigo": {
            "type": "integer",
            "format": "int32"
          },
          "id_colheita": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id_usuario_novo": 17,
          "id_usuario_antigo": 5,
          "id_colheita": 1
        }
      },
      "CadastrarColheita_VoluntarioRequest": {
        "title": "CadastrarColheita_VoluntarioRequest",
        "required": [
          "colheita",
          "voluntario"
        ],
        "type": "object",
        "properties": {
          "colheita": {
            "$ref": "#/components/schemas/Colheita"
          },
          "voluntario": {
            "$ref": "#/components/schemas/Voluntario"
          }
        },
        "example": {
          "colheita": {
            "id_colheita": 1
          },
          "voluntario": {
            "id_usuario": 17
          }
        }
      },
      "Colheita": {
        "title": "Colheita",
        "required": [
          "id_colheita"
        ],
        "type": "object",
        "properties": {
          "id_colheita": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id_colheita": 1
        }
      },
      "Voluntario": {
        "title": "Voluntario",
        "required": [
          "id_usuario"
        ],
        "type": "object",
        "properties": {
          "id_usuario": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id_usuario": 17
        }
      },
      "AtualizarColheitaRequest": {
        "title": "AtualizarColheitaRequest",
        "required": [
          "id_colheita",
          "data_colheita",
          "descricao_colheita"
        ],
        "type": "object",
        "properties": {
          "id_colheita": {
            "type": "integer",
            "format": "int32"
          },
          "data_colheita": {
            "type": "string"
          },
          "descricao_colheita": {
            "type": "string"
          }
        },
        "example": {
          "id_colheita": 12,
          "data_colheita": "2100-05-31",
          "descricao_colheita": "Descrição da colheita"
        }
      },
      "CadastrarColheitaRequest": {
        "title": "CadastrarColheitaRequest",
        "required": [
          "data_colheita",
          "descricao_colheita"
        ],
        "type": "object",
        "properties": {
          "data_colheita": {
            "type": "string"
          },
          "descricao_colheita": {
            "type": "string"
          }
        },
        "example": {
          "data_colheita": "2023-05-31",
          "descricao_colheita": "Descrição da colheita"
        }
      },
...
    }
  },
...
}